ExistsById control added before delete pinned project

+++ src/main/java/com/rte_france/antares/datamanager_back/service/impl/ProjectServiceImpl.java
@@ -1,5 +1,6 @@
 package com.rte_france.antares.datamanager_back.service.impl;
 
+import com.rte_france.antares.datamanager_back.exception.ResourceNotFoundException;
 import com.rte_france.antares.datamanager_back.repository.PinnedProjectRepository;
 import com.rte_france.antares.datamanager_back.repository.model.PinnedProjectEntity;
 import com.rte_france.antares.datamanager_back.repository.model.PinnedProjectEntityId;
@@ -29,9 +30,14 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     public void deletePinnedProjectForGivenUser(String userId, Integer projectId) {
-                PinnedProjectEntityId pinnedProjectEntityId = new PinnedProjectEntityId(userId, projectId);
-                pinnedProjectRepository.deletePinnedProjectEntityById(pinnedProjectEntityId);
-    }
+        PinnedProjectEntityId pinnedProjectEntityId = new PinnedProjectEntityId(userId, projectId);
 
+        boolean exists = pinnedProjectRepository.existsById(pinnedProjectEntityId);
+        if (!exists) {
+            throw new ResourceNotFoundException("Pinned project not found for user: " + userId + ", project ID: " + projectId);
+        }
+
+        pinnedProjectRepository.deletePinnedProjectEntityById(pinnedProjectEntityId);
+    }
 
 }
